requirements.txt
tox
PyYaml
requests
flask
tabulate
paramiko
python-bugzilla
urllib3
mako
